pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
# Use an underscore for the package name
name = "cnn_occlusion_robustness"
version = "0.1.0"
authors = [
  { name="Julian Quick" }
]
description = "A research project to study the robustness of CNNs to camera occlusion effects using the GTSRB dataset."
requires-python = ">=3.8"
license = { file="LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core dependencies for the entire project
dependencies = [
    "camera-occlusion",
    "torch",
    "torchvision",
    "numpy<2.0",
    "opencv-python",
    "imageio",
    "pyyaml",
    "tqdm",
    "pandas",
    "matplotlib",
    "seaborn",
    "scipy",
    "scikit-learn",
]

[project.urls]
"Homepage" = "https://github.com/your-username/cnn-occlusion-robustness"

[project.scripts]
train-gtsrb = "cnn_occlusion_robustness.train:main"
analyze-results = "cnn_occlusion_robustness.analysis.advanced_analysis_report:main"
visualize-filters = "cnn_occlusion_robustness.analysis.visualize_filters:main"
visualize-activations = "cnn_occlusion_robustness.analysis.visualize_activations:main"


[tool.setuptools.packages.find]
where = ["src"]
